requirements.txt
scikit-learn
pandas
numpy
streamlit
nltk
matplotlib
torch
torchvision
pillow